+++ R/tab_create_modify.R
@@ -105,7 +105,7 @@ tab_header <- function(
 #' ```r
 #' gtcars %>%
 #'   dplyr::select(
-#'     -mfr, -trim, bdy_style, drivetrain,
+#'     -mfr, -trim, bdy_style,
 #'     -drivetrain, -trsmn, -ctry_origin
 #'   ) %>%
 #'   dplyr::slice(1:8) %>%
@@ -281,8 +281,16 @@ resolve_spanner_level <- function(
   highest_level <- 0L
 
   spanners_tbl <- dplyr::select(spanners_tbl, spanner_id, vars, spanner_level)
-  highest_level <- spanners_tbl %>%
-    dplyr::filter(vapply(vars, function(x) any(column_names %in% x), logical(1))) %>%
+
+  highest_level <-
+    dplyr::filter(
+      spanners_tbl,
+      vapply(
+        vars,
+        FUN.VALUE = logical(1),
+        FUN = function(x) any(column_names %in% x)
+      )
+    ) %>%
     dplyr::pull("spanner_level") %>%
     max(0) # Max of ^ and 0
 
@@ -382,16 +390,18 @@ tab_spanner_delim <- function(
   # Perform input object validation
   stop_if_not_gt(data = data)
 
-  split <- match.arg(split)
+  split <- rlang::arg_match(split)
 
   # Get all of the columns in the dataset
   all_cols <- dt_boxhead_get_vars(data = data)
 
   # Get the columns supplied in `columns` as a character vector
   columns <-
-    resolve_cols_c(
-      expr = {{ columns }},
-      data = data
+    suppressWarnings(
+      resolve_cols_c(
+        expr = {{ columns }},
+        data = data
+      )
     )
 
   if (!is.null(columns)) {
@@ -1049,7 +1059,7 @@ tab_footnote <- function(
     placement = c("auto", "right", "left")
 ) {
 
-  placement <- match.arg(placement)
+  placement <- rlang::arg_match(placement)
 
   # Perform input object validation
   stop_if_not_gt(data = data)
@@ -1934,11 +1944,13 @@ set_style.cells_source_notes <- function(loc, data, style) {
 #' components, the subcomponents, and the element that can adjusted.
 #'
 #' @inheritParams fmt_number
-#' @param container.width,container.height The width and height of the table's
-#'   container. Can be specified as a single-length character with units of
-#'   pixels or as a percentage. If provided as a single-length numeric vector,
-#'   it is assumed that the value is given in units of pixels. The [px()] and
-#'   [pct()] helper functions can also be used to pass in numeric values and
+#' @param container.width,container.height,container.padding.x,container.padding.y
+#'   The width and height of the table's container, and, the vertical and
+#'   horizontal padding of the table's container. The container width and height
+#'   can be specified with units of pixels or as a percentage. The padding is to
+#'   be specified as a length with units of pixels. If provided as a numeric
+#'   value, it is assumed that the value is given in units of pixels. The [px()]
+#'   and [pct()] helper functions can also be used to pass in numeric values and
 #'   obtain values as pixel or percent units.
 #' @param container.overflow.x,container.overflow.y Options to enable scrolling
 #'   in the horizontal and vertical directions when the table content overflows
@@ -2256,13 +2268,15 @@ set_style.cells_source_notes <- function(loc, data, style) {
 #'
 #' @family part creation/modification functions
 #' @section Function ID:
-#' 2-11
+#' 2-12
 #'
 #' @export
 tab_options <- function(
     data,
     container.width = NULL,
     container.height = NULL,
+    container.padding.x = NULL,
+    container.padding.y = NULL,
     container.overflow.x = NULL,
     container.overflow.y = NULL,
     table.width = NULL,