+++ R/utils_render_html.R
@@ -26,7 +26,7 @@ styles_to_html <- function(styles) {
       FUN.VALUE = character(1), USE.NAMES = FALSE,
       FUN = function(x) {
         if (any(is.null(names(x)))) {
-          style <- gsub(":", ": ", x, fixed = TRUE)
+          style <- as.character(x)
         } else if (all(names(x) != "")) {
           x <- cell_style_to_html(x)
           style <- tidy_gsub(paste0(names(x), ": ", x, ";", collapse = " "), ";;", ";")
@@ -799,6 +799,7 @@ create_body_component_h <- function(data) {
   # Get vector representation of stub layout
   stub_layout <- get_stub_layout(data = data)
 
+  # Determine if there is a stub column in `stub_layout`
   has_stub_column <- "rowname" %in% stub_layout
 
   # Get a matrix of all cells in the body (not including summary cells)
@@ -1101,7 +1102,7 @@ create_body_component_h <- function(data) {
                       } else {
                         paste0(
                           " style=\"",
-                          htmltools::htmlEscape(cell_style, attribute = TRUE),
+                          htmltools::htmlEscape(cell_style, attribute = FALSE),
                           "\""
                         )
                       },
@@ -1409,11 +1410,11 @@ get_body_component_cell_matrix <- function(data) {
 
     group_label_matrix <-
       dt_stub_df_get(data = data) %>%
-      dplyr::select(-rowname, -group_label) %>%
+      dplyr::select(-row_id, -group_label) %>%
       dplyr::inner_join(groups_rows_df, by = "group_id") %>%
       dplyr::mutate(
         row = dplyr::row_number(),
-        built = dplyr::if_else(row_start != row, "", built)
+        built = dplyr::if_else(row_start != row, "", built_group_label)
       ) %>%
       dplyr::select(built) %>%
       as.matrix() %>%
@@ -1469,7 +1470,8 @@ summary_row_tags_i <- function(data, group_id) {
   summary_df <-
     dplyr::select(
       list_of_summaries$summary_df_display_list[[group_id]],
-      .env$rowname_col_private, .env$default_vars
+      dplyr::all_of(rowname_col_private),
+      dplyr::all_of(default_vars)
     )
 
   # Get effective number of columns